+++ src/table/handlers/TableHandler.ts
@@ -177,10 +177,14 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
     const accept = this.getAndCheckPayloadFormat(tableContext);
     const prefer = this.getAndCheckPreferHeader(tableContext);
 
+    // curently unable to use checking functions as the partitionKey
+    // and rowKey are not coming through the context.
+    // const partitionKey = this.getAndCheckPartitionKey(tableContext);
+    // const rowKey = this.getAndCheckRowKey(tableContext);
     if (
       !options.tableEntityProperties ||
-      !options.tableEntityProperties.PartitionKey ||
-      // rowKey may be empty string
+      // rowKey and partitionKey may be empty string
+      options.tableEntityProperties.PartitionKey === null ||
       options.tableEntityProperties.RowKey === null
     ) {
       throw StorageErrorFactory.getPropertiesNeedValue(context);