+++ ChangeLog.md
@@ -10,6 +10,7 @@ Table:
 - Corrected serialization of errors during Entity Group Transactions.
 - Correct entity tests using invalid eTag formats.
 - Added support for PATCH Verb in Table Batch Operations / Entity Group Transactions.
+- Allowed use of empty string for partitionKey and rowKey on InsertEntity.
 
 Blob: